requirements.txt
flask
pymongo
gunicorn==19.6.0
nose
PyJWT